Use base component
